requirements.txt
absl-py==2.1.0
accelerate @ git+https://github.com/huggingface/accelerate.git@31af8d4e8ee770b200b4dc00b6dde7c028232dcd
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.4
astroid==3.0.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.30.0
attrs==23.1.0
beautifulsoup4==4.12.2
bitsandbytes==0.42.0
blinker==1.7.0
bokeh==3.7.3
cattrs==24.1.3
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
databases==0.8.0
datasets==3.0.1
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
executing==2.0.1
faiss-cpu==1.11.0
fastapi==0.115.8
fastf1==3.5.3
filelock==3.16.1
Flask==3.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fpdf==1.7.2
frozendict==2.3.10
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.2
httptools==0.6.4
httpx==0.25.1
huggingface-hub==0.33.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.7
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
keras==3.6.0
kiwisolver==1.4.7
langchain==0.3.26
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.5
lazy-object-proxy==1.9.0
libclang==18.1.1
lxml==4.9.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
namex==0.0.8
narwhals==1.42.1
nest-asyncio==1.6.0
networkx==3.4.1
numpy==1.26.2
openai==1.90.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.18
packaging==23.2
pandas==2.1.4
parso==0.8.4
peewee==3.17.0
peft @ git+https://github.com/huggingface/peft.git@5fe7f8f8abe914d313fc3751f2ea92de7718fbaf
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.0.0
pluggy==1.3.0
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pylance==0.16.1
pylint==3.0.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==26.2.0
rank-bm25==0.2.2
RapidFuzz==3.12.2
regex==2024.9.11
requests==2.32.4
requests-cache==1.2.1
requests-toolbelt==1.0.0
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tiktoken==0.9.0
timple==0.1.8
tokenizers==0.21.1
tomlkit==0.12.3
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@2166b6b4ff09f6dd3867ab982f262f66482aa968
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.21.0
wcwidth==0.2.12
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
wheel==0.44.0
widgetsnbextension==4.0.14
wrapt==1.16.0
xxhash==3.5.0
xyzservices==2025.4.0
yarl==1.15.5
yfinance==0.2.33
zstandard==0.23.0
